+++ packages/sdk/src/types/nft/burn/domain.ts
@@ -1,6 +1,7 @@
 import type { ItemId } from "@rarible/api-client"
 import type { IBlockchainTransaction } from "@rarible/sdk-transaction"
 import type { BigNumber } from "@rarible/types/build/big-number"
+import type { Creator } from "@rarible/api-client/build/models/Creator"
 import type { AbstractPrepareResponse } from "../../../common/domain"
 
 export type PrepareBurnRequest = {
@@ -12,6 +13,10 @@ export type BurnRequest = {
    * Number of NFTs to transfer
    */
 	amount?: number
+	/**
+   * Item creators
+   */
+	creators?: Creator[]
 } | void
 
 export interface PrepareBurnResponse extends AbstractPrepareResponse<"burn", BurnRequest, IBlockchainTransaction | void>{